third_party/WebKit/LayoutTests/platform/win/media/encrypted-media/encrypted-media-unique-origin-expected.txt

This is a testharness.js-based test.
PASS requestMediaKeySystemAccess() is not available on unique origin 
Harness: the test ran to completion.
